% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{coef.fuzzydid}
\alias{coef.fuzzydid}
\title{coef.fuzzydid}
\usage{
\method{coef}{fuzzydid}(object, ...)
}
\arguments{
\item{object}{A fuzzydid object.}

\item{...}{Unused.}
}
\value{
A named numeric vector of LATE and LQTE point estimates.
}
\description{
Extract fuzzydid point estimates.
}
\examples{
df <- expand.grid(i = seq_len(20), g = 0:1, t = 0:1)
df$d <- as.integer(df$i <= c(4, 8, 6, 16)[1 + df$t + 2 * df$g])
df$y <- 1 + 0.5 * df$g + 0.4 * df$t + 2 * df$d + sin(df$i / 7)

fit <- fuzzydid(df, y ~ d, group = "g", time = "t", did = TRUE, nose = TRUE)
coef(fit)
}
