% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{plot.fuzzydid}
\alias{plot.fuzzydid}
\title{plot.fuzzydid}
\usage{
\method{plot}{fuzzydid}(x, ...)
}
\arguments{
\item{x}{A fuzzydid object.}

\item{...}{Unused.}
}
\value{
The input \code{x}, returned invisibly.
}
\description{
Plot fuzzydid point estimates and stored confidence intervals
as a base R dot-and-whisker plot.
}
\examples{
df <- expand.grid(i = seq_len(20), g = 0:1, t = 0:1)
df$d <- as.integer(df$i <= c(4, 8, 6, 16)[1 + df$t + 2 * df$g])
df$y <- 1 + 0.5 * df$g + 0.4 * df$t + 2 * df$d + sin(df$i / 7)

fit <- fuzzydid(df, y ~ d, group = "g", time = "t", did = TRUE, nose = TRUE)
plot(fit)
}
