% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{wald_did}
\alias{wald_did}
\title{wald_did}
\usage{
wald_did(i, df, y_name, g_name, t_name, d_name, X_name = NULL)
}
\arguments{
\item{i}{Rows of the data frame to operate on}

\item{df}{Data frame to operate on}

\item{y_name}{Outcome variable}

\item{g_name}{Group variable}

\item{t_name}{Time period variable}

\item{d_name}{Treatment variable, can be any ordered variable}

\item{X_name}{Covariates to include}
}
\description{
Computes the Wald-DiD estimator, equivalent to an IV regression of y on d with
g and t as included instruments and g:t as the excluded instrument for d
}
