% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{fuzzydid}
\alias{fuzzydid}
\title{fuzzydid}
\usage{
fuzzydid(
  df,
  y_name,
  g_name,
  t_name,
  d_name,
  X_name = NULL,
  est = c("did", "tc", "cic"),
  nboot = 50
)
}
\arguments{
\item{df}{Data frame to operate on}

\item{y_name}{Outcome variable}

\item{g_name}{Group variable}

\item{t_name}{Time period variable}

\item{d_name}{Treatment variable, can be any ordered variable}

\item{X_name}{Covariates to include}

\item{est}{Estimators to compute. Should contain at least one of "did", "tc", "cic"}

\item{nboot}{Number of bootstrap samples for standard errors}
}
\description{
Computes the corrected LATE estimators as in de Chaisemartin and D'Haultfoeuille (2018)
<doi:10.1093/restud/rdx049>
}
