% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{glance.fuzzydid}
\alias{glance.fuzzydid}
\title{glance.fuzzydid}
\usage{
\method{glance}{fuzzydid}(x, ...)
}
\arguments{
\item{x}{A fuzzydid object}

\item{...}{Extra arguments are ignored}
}
\description{
glance for modelsummary
}
