% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{summary.fuzzydid}
\alias{summary.fuzzydid}
\title{summary.fuzzydid}
\usage{
\method{summary}{fuzzydid}(object, ...)
}
\arguments{
\item{object}{A fuzzydid object}

\item{...}{Extra arguments are ignored}
}
\description{
Displays the estimations in a table
}
