% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfuzzydid.R
\name{tidy.fuzzydid}
\alias{tidy.fuzzydid}
\title{tidy.fuzzydid}
\usage{
\method{tidy}{fuzzydid}(x, ...)
}
\arguments{
\item{x}{A fuzzydid object.}

\item{...}{Unused.}
}
\description{
Tidy extractor for `fuzzydid` objects.
}
